op-batcher: Split batch_submitter.go into several files
This simply moves code around into new files without changing it.
The intention is to tease it apart, but that is not done yet.
